Update hash_test.rb
There has been a lot of confusion about some of these hash tests, both in my cohort and the current module 1. The problem is that saying something like `assert_equal {"apple" => "a", "banana" => "b"}, {"a" => "apple", "b" => "banana"}.invert` doesn't work, because Ruby thinks that the curly bracket is a block instead of a hash. There are a couple ways to make this more clear for beginners, but personally I think this way works the best. Maybe this will help out 1510 when they work on this.